+++ Models/Lyric.cs
@@ -6,7 +6,7 @@ namespace Lyrics.Models;
 [JsonConverter(typeof(LyricConverter))]
 public interface ILyric
 {
-    int LyricId { get; set; }
+    long LyricId { get; set; }
     int StartTime { get; set; }
     string VideoId { get; set; }
     string Title { get; set; }
@@ -17,7 +17,7 @@ public class Lyric : ILyric
 {
     public string VideoId { get; set; } = "";
     public int StartTime { get; set; }
-    public int LyricId { get; set; }
+    public long LyricId { get; set; }
     public string Title { get; set; } = "";
     public float Offset { get; set; }
 }
